Fix comma in JSON resp
