+++ internal/pkg/common/json.go
@@ -1,5 +1,7 @@
 package common
 
+import "encoding/json"
+
 func MakeJSONArray(items [][]byte) []byte {
 	var jsonBody = []byte{'['}
 	for idx, item := range items {
@@ -10,4 +12,16 @@ func MakeJSONArray(items [][]byte) []byte {
 	}
 	jsonBody = append(jsonBody, []byte{']'}...)
 	return jsonBody
+}
+
+func UnionToJSON(items ...interface{}) (response []byte) {
+	for _, item := range items {
+		body, err := json.Marshal(item)
+		if err != nil {
+			continue
+		}
+		response = append(response, body...)
+		response = append(response, byte(','))
+	}
+	return
 }
